+++ lua/plugins/config/nvim_gomove.lua
@@ -1,0 +1,8 @@
+return function(use)
+	use({
+		'booperlv/nvim-gomove',
+		config = function()
+			require('gomove').setup({})
+		end,
+	})
+end