+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Feat
+- update
 - moved to coq and chadtree
 - **airline:** removed file icon in explorer tree in bottom right
 - **airline:** added support to show the tree plugin name