+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 ## [Unreleased]
 
 ### Feat
+- moved to coq and chadtree
 - removed plenary from lazygit plugin
 - changed lazygit to direct integration
 - update
-- moved to coq and chadtree
 - **airline:** removed file icon in explorer tree in bottom right
 - **airline:** added support to show the tree plugin name
+- **nvim-tree:** added `.git` to ignore list
 
 ### Revert
 - return to nvim tree because chadtree theme is ugly